// packages/deciphony-renderer/src/standardStaff/enums/standardStaffSkinKeyEnum.d.ts
export declare enum StandardStaffSkinKeyEnum {
    Measure = "measure",
    Treble = "treble",
    Alto = "alto",
    Bass = "bass",
    Tenor = "Tenor",
    Treble_f = "treble_f",
    Alto_f = "alto_f",
    Bass_f = "bass_f",
    Tenor_f = "tenor_f",
    Sharp = "sharp",
    Flat = "flat",
    Double_sharp = "double_sharp",
    Double_flat = "double_flat",
    Natural = "natural",
    Close_line = "close_line",
    linked_close_line = "linked_close_line",
    Single_barline = "single_barline",
    Double_barline = "double_barline",
    StartRepeat_barline = "startRepeat_barline",
    EndRepeat_barline = "endRepeat_barline",
    Dashed_barline = "dashed_barline",
    Final_barline = "final_barline",
    Start_end_repeat_barline = "start_end_repeat_barline",
    Dotted_barline = "dotted_barline",
    Reverse_barline = "reverse_barline",
    Heavy_barline = "heavy_barline",
    Heavy_double_barline = "heavy_double_barline",
    linked_single_barline = "linked_single_barline",
    linked_double_barline = "linked_double_barline",
    linked_startRepeat_barline = "linked_startRepeat_barline",
    linked_endRepeat_barline = "linked_endRepeat_barline",
    linked_dashed_barline = "linked_dashed_barline",
    linked_final_barline = "linked_final_barline",
    linked_start_end_repeat_barline = "linked_start_end_repeat_barline",
    linked_dotted_barline = "linked_dotted_barline",
    linked_reverse_barline = "linked_reverse_barline",
    linked_heavy_barline = "linked_heavy_barline",
    linked_heavy_double_barline = "linked_heavy_double_barline",
    C = "C",
    C_sharp = "C_sharp",
    D_flat = "D_flat",
    D = "D",
    E_flat = "E_flat",
    E = "E",
    F = "F",
    F_sharp = "F_sharp",
    G_flat = "G_flat",
    G = "G",
    A_flat = "A_flat",
    A = "A",
    B_flat = "B_flat",
    B = "B",
    C_flat = "C_flat",
    '1_1' = "1_1",
    '1_4' = "1_4",
    '2_4' = "2_4",
    '3_4' = "3_4",
    '4_4' = "4_4",
    '3_8' = "3_8",
    '6_8' = "6_8",
    NoteHead_1 = "noteHead_1",
    NoteHead_2 = "noteHead_2",
    NoteHead_3 = "noteHead_3",
    NoteStem = "notestem",
    NoteTail_1 = "noteTail_1",
    NoteTail_2 = "noteTail_2",
    NoteTail_3 = "noteTail_3",
    NoteTail_4 = "noteTail_4",
    NoteTail_5 = "noteTail_5",
    NoteTail_6 = "noteTail_6",
    NoteTail_1_r = "noteTail_1_r",
    NoteTail_2_r = "noteTail_2_r",
    NoteTail_3_r = "noteTail_3_r",
    NoteTail_4_r = "noteTail_4_r",
    NoteTail_5_r = "noteTail_5_r",
    NoteTail_6_r = "noteTail_6_r",
    Rest_1 = "rest_1",
    Rest_2 = "rest_2",
    Rest_3 = "rest_3",
    Rest_4 = "rest_4",
    Rest_5 = "rest_5",
    Rest_6 = "rest_6",
    Rest_7 = "rest_7",
    Rest_8 = "rest_8",
    Rest_9 = "rest_9",
    AugmentationDot_1 = "augmentationDot_1",
    AugmentationDot_2 = "augmentationDot_2",
    AugmentationDot_3 = "augmentationDot_3",
    AddLine_u = "addLine_u",// 上方的加线
    AddLine_d = "addLine_d",// 下方的加线
    Bracket = "bracket",// 标准连谱号
    Brace = "brace",// 大括号
    Square = "square",// 中括号
    Repeat_coda = "repeat_coda",
    Repeat_to_coda = "repeat_to_coda",
    Repeat_segno = "repeat_segno",
    Repeat_dc = "repeat_dc",
    Repeat_ds = "repeat_ds",
    Repeat_fine = "repeat_fine",
    Repeat_dc_al_fine = "repeat_dc_al_fine",
    Repeat_dc_al_coda = "repeat_dc_al_coda",
    Repeat_ds_al_fine = "repeat_ds_al_fine",
    Repeat_ds_al_coda = "repeat_ds_al_coda",
    Accent_above = "acccent_above",
    Staccato_above = "staccato_above",
    Staccatissimo_above = "staccato_above",
    Tenuto_above = "tenuto_above",
    Loure_above = "loure_above",
    Marcato_above = "marcato_above",
    Accent_staccato_above = "accent_staccato_above",
    Marcato_staccato_above = "marcato_staccato_above",
    Marcato_tenuto_above = "marcato_tenuto_above",
    Staccatissimo_stroke_above = "staccatissimo_stroke_above",
    Staccatissimo_wedge_above = "staccatissimo_wedge_above",
    Stress_above = "stress_above",
    Tenuto_accent_above = "tenuto_accent_above",
    Unstress_above = "unstress_above",
    Open = "open",
    Muted = "muted",
    Harmonic = "harmonic",
    Up_bow = "up_bow",
    Down_bow = "down_bow",
    Turn = "turn",
    InvertedTurn = "invertedTurn",
    Trill = "trill",
    ShortTrill = "shortTrill",
    Mordent = "mordent",
    Arpeggio = "arpeggio",
    Up_arpeggio = "up_arpeggio",
    Down_arpeggio = "down_arpeggio",
    Bracket_arpeggio = "bracket_arpeggio",
    Up_arpeggio_straight = "up_arpeggio_straight",
    Down_arpeggio_straight = "down_arpeggio_straight",
    Dyn_PPP = "dyn_ppp",
    Dyn_PP = "dyn_pp",
    Dyn_P = "dyn_p",
    Dyn_MP = "dyn_mp",
    Dyn_MF = "dyn_mf",
    Dyn_F = "dyn_f",
    Dyn_FF = "dyn_ff",
    Dyn_FFF = "dyn_fff",
    Dyn_FP = "dyn_fp",
    Dyn_PF = "dyn_pf",
    Dyn_SF = "dyn_sf",
    Dyn_SFZ = "dyn_sfz",
    Dyn_SFF = "dyn_sff",
    Dyn_SFFZ = "dyn_sffz",
    Dyn_SFP = "dyn_sfp",
    Dyn_RFZ = "dyn_rfz",
    Dyn_RF = "dyn_rf",
    Dyn_FZ = "dyn_fz"
}